Add bean validation for post

+++ src/main/kotlin/com/junior/blog/controller/PostController.kt
@@ -1,5 +1,6 @@
 package com.junior.blog.controller
 
+import com.junior.blog.controller.util.getErrors
 import com.junior.blog.model.Post
 import com.junior.blog.model.User
 import com.junior.blog.service.CategoryServiceImpl
@@ -9,7 +10,9 @@ import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
+import org.springframework.validation.BindingResult
 import org.springframework.web.bind.annotation.*
+import javax.validation.Valid
 
 @Controller
 @ControllerAdvice
@@ -23,8 +26,18 @@ class PostController(
     fun postForm(model: Model) = "add-post"
     
     @PostMapping("new/add")
-    fun createPost(post: Post,
-                   @AuthenticationPrincipal user: User): String {
+    fun createPost(@AuthenticationPrincipal user: User,
+                   @Valid post: Post,
+                   bindRes: BindingResult,
+                   model: Model
+    ): String {
+        if (bindRes.hasErrors()) {
+            model.mergeAttributes(getErrors(bindRes))
+            model.addAttribute("post", post)
+            
+            return "add-post"
+        }
+        
         post.user = user
         postService.save(post)
         categoryService.deleteAll(categoryService.getAllEmpty())
